vulnerablility issue fix

+++ Myrtille.Services/FileStorage.cs
@@ -27,6 +27,7 @@ namespace Myrtille.Services
 {
     public class FileStorage : IFileStorage
     {
+
         public List<string> GetUserDocumentsFolderFiles(
             Guid remoteSessionId,
             string userDomain,
@@ -34,57 +35,78 @@ namespace Myrtille.Services
             string userPassword,
             string sharedFolderPath)
         {
-            var documentsFolder = sharedFolderPath;
+            string installDir = AppDomain.CurrentDomain.BaseDirectory;
+            //var documentsFolder = Path.Combine(
+            //    installDir,
+            //    @"..\..\share_rdp_folder", sharedFolderPath
+            //);
 
-            try
+            string shareRDPPath = @"..\..\share_rdp_folder";
+            var documentsFolder = Path.Combine(
+                installDir, shareRDPPath
+            );
+            documentsFolder += "\\" + sharedFolderPath + "\\";
+            string finalFullPath = Path.GetFullPath(documentsFolder);
+            if (finalFullPath.StartsWith(Path.GetFullPath(Path.Combine(installDir, shareRDPPath))))
             {
-                var fileNames = Directory.GetFiles(documentsFolder);
-                for (var i = 0; i < fileNames.Length; i++)
+                try
                 {
-                    fileNames[i] = Path.GetFileName(fileNames[i]);
+                    var fileNames = Directory.GetFiles(documentsFolder);
+                    for (var i = 0; i < fileNames.Length; i++)
+                    {
+                        fileNames[i] = Path.GetFileName(fileNames[i]);
+                    }
+                    return new List<string>(fileNames);
                 }
-                return new List<string>(fileNames);
+                catch (Exception exc)
+                {
+                    Trace.TraceError("Failed to retrieve file(s) from user {0} documents folder {1}, remote session {2} ({3})", userName, documentsFolder, remoteSessionId, exc);
+                    throw;
+                }
             }
-            catch (Exception exc)
-            {
-                Trace.TraceError("Failed to retrieve file(s) from user {0} documents folder {1}, remote session {2} ({3})", userName, documentsFolder, remoteSessionId, exc);
-                throw;
-            }
+            return new List<string>();
         }
-        
+
         public void UploadFileToUserDocumentsFolder(
             UploadRequest uploadRequest)
         {
-            var documentsFolder = uploadRequest.SharedFolderPath;
-
-            try
+            string installDir = AppDomain.CurrentDomain.BaseDirectory;
+            string shareRDPPath = @"..\..\share_rdp_folder";
+            var documentsFolder = Path.Combine(
+                installDir, shareRDPPath
+            );
+            documentsFolder += "\\" + uploadRequest.SharedFolderPath + "\\";
+            string finalFullPath = Path.GetFullPath(documentsFolder);
+            if (finalFullPath.StartsWith(Path.GetFullPath(Path.Combine(installDir, shareRDPPath))))
             {
-                // replace if already exists
-                if (File.Exists(Path.Combine(documentsFolder, uploadRequest.FileName)))
+                try
                 {
-                    File.Delete(Path.Combine(documentsFolder, uploadRequest.FileName));
+                    if (File.Exists(Path.Combine(documentsFolder, uploadRequest.FileName)))
+                    {
+                        File.Delete(Path.Combine(documentsFolder, uploadRequest.FileName));
+                    }
+
+                    var fileStream = File.Create(Path.Combine(documentsFolder, uploadRequest.FileName));
+
+                    int bytesRead;
+                    var buffer = new byte[4096];
+
+                    while ((bytesRead = uploadRequest.FileStream.Read(buffer, 0, buffer.Length)) > 0)
+                    {
+                        fileStream.Write(buffer, 0, bytesRead);
+                    }
+
+                    fileStream.Close();
+                    uploadRequest.FileStream.Close();
                 }
-
-                var fileStream = File.Create(Path.Combine(documentsFolder, uploadRequest.FileName));
-
-                int bytesRead;
-                var buffer = new byte[4096];
-
-                while ((bytesRead = uploadRequest.FileStream.Read(buffer, 0, buffer.Length)) > 0)
+                catch (Exception exc)
                 {
-                    fileStream.Write(buffer, 0, bytesRead);
+                    Trace.TraceError("Failed to upload file {0} to user {1} documents folder {2}, remote session {3} ({4})", uploadRequest.FileName, uploadRequest.UserName, documentsFolder, uploadRequest.RemoteSessionId, exc);
+                    throw;
                 }
-
-                fileStream.Close();
-                uploadRequest.FileStream.Close();
-            }
-            catch (Exception exc)
-            {
-                Trace.TraceError("Failed to upload file {0} to user {1} documents folder {2}, remote session {3} ({4})", uploadRequest.FileName, uploadRequest.UserName, documentsFolder, uploadRequest.RemoteSessionId, exc);
-                throw;
             }
 
-            Trace.TraceInformation("Uploaded file {0} to user {1} documents folder {2}, remote session {3}", uploadRequest.FileName, uploadRequest.UserName, documentsFolder, uploadRequest.RemoteSessionId);
+            Trace.TraceInformation("Uploaded file {0} to user {1} documents folder {2}, remote session {3}", uploadRequest.FileName, uploadRequest.UserName, uploadRequest.RemoteSessionId);
         }
 
         public Stream DownloadFileFromUserDocumentsFolder(
@@ -95,20 +117,29 @@ namespace Myrtille.Services
             string fileName,
             string sharedFolderPath)
         {
-            var documentsFolder = sharedFolderPath;
-
             Stream fileStream = null;
 
-            try
+            string installDir = AppDomain.CurrentDomain.BaseDirectory;
+            string shareRDPPath = @"..\..\share_rdp_folder";
+            var documentsFolder = Path.Combine(
+                installDir, shareRDPPath
+            );
+            documentsFolder += "\\" + sharedFolderPath + "\\";
+            string finalFullPath = Path.GetFullPath(documentsFolder);
+            if (finalFullPath.StartsWith(Path.GetFullPath(Path.Combine(installDir, shareRDPPath))))
             {
-                fileStream = File.Open(Path.Combine(documentsFolder, fileName), FileMode.Open, FileAccess.Read, FileShare.Read);
-            }
-            catch (Exception exc)
-            {
-                Trace.TraceError("Failed to download file {0} from user {1} documents folder {2}, remote session {3} ({4})", fileName, userName, documentsFolder, remoteSessionId, exc);
-                throw;
-            }
 
+                try
+                {
+                    fileStream = File.Open(Path.Combine(documentsFolder, fileName), FileMode.Open, FileAccess.Read, FileShare.Read);
+                }
+                catch (Exception exc)
+                {
+                    Trace.TraceError("Failed to download file {0} from user {1} documents folder {2}, remote session {3} ({4})", fileName, userName, documentsFolder, remoteSessionId, exc);
+                    throw;
+                }
+
+            }
             Trace.TraceInformation("Downloaded file {0} from user {1} documents folder {2}, remote session {3}", fileName, userName, documentsFolder, remoteSessionId);
 
             return fileStream;
